Do everything in worker controller.
